added search functionality, login and favorites functionality features

+++ app/src/main/java/com/example/weatherapplication/data/model/CurrentConditionsResponse/CurrentConditionsItem.kt
@@ -27,7 +27,7 @@ data class CurrentConditionsItem(
     val UVIndex: Double,
     val UVIndexText: String,
     val Visibility: Visibility,
-    val WeatherIcon: Double,
+    val WeatherIcon: Int,
     val WeatherText: String,
     val WetBulbGlobeTemperature: WetBulbGlobeTemperature,
     val WetBulbTemperature: WetBulbTemperature,
