+++ app/src/main/java/com/example/weatherapplication/data/model/DailyWeatherForecast/RelativeHumidity.kt
@@ -1,7 +1,7 @@
 package com.example.weatherapplication.data.model.DailyWeatherForecast
 
 data class RelativeHumidity(
-    val Average: Int,
-    val Maximum: Int,
-    val Minimum: Int
+    val Average: Double,
+    val Maximum: Double,
+    val Minimum: Double
 )
