+++ app/src/main/java/com/example/weatherapplication/data/model/DailyWeatherForecast/Snow.kt
@@ -2,6 +2,6 @@ package com.example.weatherapplication.data.model.DailyWeatherForecast
 
 data class Snow(
     val Unit: String,
-    val UnitType: Int,
-    val Value: Int
+    val UnitType: Double,
+    val Value: Double
 )
