+++ app/src/main/java/com/example/weatherapplication/data/model/DailyWeatherForecast/SolarIrradiance.kt
@@ -2,6 +2,6 @@ package com.example.weatherapplication.data.model.DailyWeatherForecast
 
 data class SolarIrradiance(
     val Unit: String,
-    val UnitType: Int,
+    val UnitType: Double,
     val Value: Double
 )
