+++ app/src/main/java/com/example/weatherapplication/data/model/DailyWeatherForecast/TotalLiquid.kt
@@ -2,6 +2,6 @@ package com.example.weatherapplication.data.model.DailyWeatherForecast
 
 data class TotalLiquid(
     val Unit: String,
-    val UnitType: Int,
+    val UnitType: Double,
     val Value: Double
 )
