+++ app/src/main/java/com/example/weatherapplication/data/model/HourlyData/Ceiling.kt
@@ -2,6 +2,6 @@ package com.example.weatherapplication.data.model.HourlyData
 
 data class Ceiling(
     val Unit: String,
-    val UnitType: Int,
-    val Value: Int
+    val UnitType: Double,
+    val Value: Double
 )
