+++ app/src/main/java/com/example/weatherapplication/data/model/HourlyData/Direction.kt
@@ -1,7 +1,7 @@
 package com.example.weatherapplication.data.model.HourlyData
 
 data class Direction(
-    val Degrees: Int,
+    val Degrees: Double,
     val English: String,
     val Localized: String
 )
