+++ app/src/main/java/com/example/weatherapplication/data/model/HourlyData/RealFeelTemperatureShade.kt
@@ -3,6 +3,6 @@ package com.example.weatherapplication.data.model.HourlyData
 data class RealFeelTemperatureShade(
     val Phrase: String,
     val Unit: String,
-    val UnitType: Int,
+    val UnitType: Double,
     val Value: Double
 )
