+++ app/src/main/java/com/example/weatherapplication/data/model/HourlyData/Temperature.kt
@@ -2,6 +2,6 @@ package com.example.weatherapplication.data.model.HourlyData
 
 data class Temperature(
     val Unit: String,
-    val UnitType: Int,
+    val UnitType: Double,
     val Value: Double
 )
